+++ src/ios/PlayGround/PlayGround/Presentation/CreateTab/Upload/Progress.swift
@@ -1,0 +1,14 @@
+//
+//  Progress.swift
+//  PlayGround
+//
+//  Created by chuiseo-MN on 2022/02/08.
+//
+
+import Foundation
+
+enum Progress {
+    case inProgress
+    case fail
+    case success
+}
