+++ src/ios/PlayGround/PlayGround/Presentation/CreateTab/Upload/UploadProgress.swift
@@ -1,0 +1,13 @@
+//
+//  UploadProgress.swift
+//  PlayGround
+//
+//  Created by chuiseo-MN on 2022/02/08.
+//
+
+import Foundation
+
+class UploadProgress {
+    static let shared = UploadProgress()
+    @Published var isFinished: Progress = .inProgress
+}
