+++ src/ios/PlayGround/PlayGround/Presentation/CreateTab/Upload/UploadViewController.swift
@@ -383,32 +383,3 @@ extension UploadViewController: UITextViewDelegate {
         }
     }
 }
-
-class UploadProgress {
-    static let shared = UploadProgress()
-    @Published var isFinished: Progress = .inProgress
-}
-
-enum Progress {
-    case inProgress
-    case fail
-    case success
-}
-
-extension UIViewController {
-    static func showToastMessage(toastLabel: UILabel, font: UIFont = UIFont.systemFont(ofSize: 12, weight: .light)) {
-        let window = UIApplication.shared.windows.first!
-        toastLabel.frame = CGRect(x: window.frame.width / 2 - 150, y: window.frame.height - 180, width: 300, height: 50)
-        
-        toastLabel.backgroundColor = UIColor.black.withAlphaComponent(0.7)
-        toastLabel.textColor = UIColor.white
-        toastLabel.numberOfLines = 2
-        toastLabel.font = font
-        toastLabel.textAlignment = .center
-        toastLabel.layer.cornerRadius = 25
-        toastLabel.clipsToBounds = true
-        
-        guard let topController = UIApplication.shared.keyWindow?.rootViewController else { return }
-        topController.view.addSubview(toastLabel)
-    }
-}
